hasil migrasi kode php - part 8

+++ UnpakSipaksi.Modules.PenelitianHibah.Application/GetRab/RabResponse.cs
@@ -10,9 +10,9 @@ namespace UnpakSipaksi.Modules.PenelitianHibah.Application.GetRab
     {
         public string? Uuid { get; set; }
         public string? UuidPenelitianHibah { get; set; }
-        public int? HargaSatuan { get; set; }
-        public int? Item { get; set; }
-        public int? Total { get; set; }
+        public long? HargaSatuan { get; set; }
+        public long? Item { get; set; }
+        public long? Total { get; set; }
         public KelompokResponse? Kelompok { get; set; }
         public KomponenResponse? Komponen { get; set; }
         public SatuanResponse? Satuan { get; set; }
